temp_sv/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==2.3.4
httpx==0.26.0
cryptography==41.0.7
python-dotenv==1.0.0
jinja2==3.1.3
numpy==1.26.3
